[refactor] Move netty setup into separate class
